renamed NewDatePicker to DatePicker
