+++ src/stories/Form.stories.js
@@ -2,6 +2,7 @@ import React, {useState} from 'react';
 import {View} from 'react-native';
 import AddressSearch from '@components/AddressSearch';
 import CheckboxWithLabel from '@components/CheckboxWithLabel';
+import DatePicker from '@components/DatePicker';
 import FormProvider from '@components/Form/FormProvider';
 import InputWrapper from '@components/Form/InputWrapper';
 import Picker from '@components/Picker';
@@ -29,6 +30,7 @@ const story = {
         CheckboxWithLabel,
         Picker,
         StatePicker,
+        DatePicker,
     },
 };
 
@@ -66,6 +68,11 @@ function Template(args) {
                 inputID="street"
                 containerStyles={[styles.mt4]}
                 hint="No PO box"
+            />
+            <DatePicker 
+                inputID="dob"
+                label="Date of Birth"
+                containerStyles={[styles.mt4]}
             />
             <View>
                 <InputWrapper
